Kaz experiments (#3)
* Worked on the landing page, nav bar and the inception to a new dashboard

* Added light/dark themes, refactored the slices and layouts, and updated the login prompt.

* Updated the storemap due to added Leaflet map component

* Updates for Leaflet again

* Homepage is now more dynamic. Both Auth scenarios kinda work but need to be refactored.

* Refactored the Auth component, worked on making the theme work across components, prepared to delete a lot of duplicate or redundant files (named TBD for To Be Deleted). Ensured that the auth buttons only render for non-authenticated users.

* Added Welcome page and related components

* Update README.md

* Update README again.md

* Updated readme, added basic Welcome page components

* Deleted old useAuth hook, no longer used.

* Resolved merge conflicts in lock files

* Got the cubid uid and authentication working, saves to supabase, and started cleaning up the welcome flow.

* Kept working on getting & storing the Cubid data

* Got the welcome carousel flowing a little better, updated themes for the steps, and reduced from 10 to 6 steps.

* Deleted the original Protected page from Cubid Starter

* Cleaned up the data handling in the welcome flow

* fix vercel build

---------

+++ components/NavProfileLinks.tsx
@@ -44,11 +44,14 @@ function NavProfileLinks() {
   };
 
   const logoutUser = async () => {
+    // Optional: Make the logout request to the backend
     // await axios.get(process.env.NEXT_PUBLIC_BASE_URL+'/user/logout')
-    localStorage.clear();
-    window.location = "/";
+    if (typeof window !== 'undefined') {
+      localStorage.clear();
+      window.location.href = "/";
+    }
   };
-
+  
   const handleDropDownClick = () => {
     const elem = document.activeElement;
     if (elem) {
